+++ bindings/hyperdrivepy/src/hyperdrive_state_methods/short.rs
@@ -1,3 +1,3 @@
-// mod close;
+mod close;
 // mod max;
 // mod open;
